This commit was manufactured by cvs2svn to create tag 'LIGHTNING_0_3_1_RELEASE'.

+++ mozilla/nsprpub/pr/src/md/windows/w32rng.c
@@ -59,8 +59,8 @@ CurrentClockTickTime(LPDWORD lpdwHigh, LPDWORD lpdwLow)
 extern PRSize _PR_MD_GetRandomNoise( void *buf, PRSize size )
 {
     DWORD   dwHigh, dwLow, dwVal;
-    size_t  n = 0;
-    size_t  nBytes;
+    int     n = 0;
+    int     nBytes;
     time_t  sTime;
 
     if (size <= 0)
